+++ doc/release-notes.md
@@ -4,3 +4,4 @@ Bitcoin ABC version 0.17.2 is now available from:
 
 This release includes the following features and fixes:
  - Remove deprecated `estimatepriority` RPC.
+ - Remove deprecated `estimatesmartpriority` RPC.